+++ modules/contrib/administerusersbyrole/tests/src/Functional/AdministerusersbyroleTest.php
@@ -3,218 +3,250 @@
 namespace Drupal\Tests\administerusersbyrole\Functional;
 
 use Drupal\Tests\BrowserTestBase;
-use Drupal\Component\Utility\SafeMarkup;
-use Drupal\Core\Session\AccountInterface;
 
 /**
- * Testing for administerusersbyrole module
+ * Testing for administerusersbyrole module.
  *
  * @group administerusersbyrole
  */
 class AdministerusersbyroleTest extends BrowserTestBase {
 
-  public static $modules = array('administerusersbyrole', 'user');
+  /**
+   * Modules to install.
+   *
+   * @var array
+   */
+  public static $modules = ['administerusersbyrole', 'user'];
 
-  protected $roles = array();
-  protected $users = array();
+  /**
+   * {@inheritdoc}
+   */
+  protected $defaultTheme = 'classy';
 
-  public function setUp() {
+  protected $roles = [];
+  protected $users = [];
+
+  /**
+   * The access manager.
+   *
+   * @var \Drupal\administerusersbyrole\Services\AccessManagerInterface
+   */
+  protected $accessManager;
+
+  /**
+   * Editable module configuration.
+   */
+  protected $config;
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function setUp() {
     parent::setUp();
+    $this->accessManager = \Drupal::service('administerusersbyrole.access');
+    $this->config = \Drupal::service('config.factory')->getEditable('administerusersbyrole.settings');
 
-    $this->createUserWithRole('noroles', array());
+    $this->createUserWithRole('noroles', []);
     $this->createRolesAndUsers('alpha', FALSE);
     $this->createRolesAndUsers('beta', TRUE);
-    $this->createUserWithRole('alphabeta', array('alpha', 'beta'));
+    $this->createUserWithRole('alphabeta', ['alpha', 'beta']);
 
-    // alphabeta_ed
-    $perms = array(
+    // alphabeta_ed.
+    $perms = [
       'access content',
-      _administerusersbyrole_build_perm_string($this->roles['alpha'], 'edit'),
-      _administerusersbyrole_build_perm_string($this->roles['beta'], 'edit'),
-    );
-    $this->roles['alphabeta_ed'] = $this->drupalCreateRole($perms, 'alphabeta_ed');
-    $this->createUserWithRole('alphabeta_ed', array('alphabeta_ed'));
+      $this->accessManager->buildPermString('edit'),
+      $this->accessManager->buildPermString('edit', 'alpha'),
+      $this->accessManager->buildPermString('edit', 'beta'),
+    ];
+    $this->drupalCreateRole($perms, 'alphabeta_ed');
+    $this->config->set("roles.alphabeta_ed", 'perm')->save();
+    $this->createUserWithRole('alphabeta_ed', ['alphabeta_ed']);
 
-    // all_editor
-    $perms = array(
+    // all_editor.
+    $perms = [
       'access content',
-      _administerusersbyrole_build_perm_string(AccountInterface::AUTHENTICATED_ROLE, 'edit'),
-    );
-    foreach ($this->roles as $roleName => $roleID) {
-      $perms[] = _administerusersbyrole_build_perm_string($this->roles[$roleName], 'edit');
+      $this->accessManager->buildPermString('edit'),
+    ];
+    foreach (array_keys($this->accessManager->managedRoles()) as $roleName) {
+      $perms[] = $this->accessManager->buildPermString('edit', $roleName);
     }
-    $this->roles['all_editor'] = $this->drupalCreateRole($perms, 'all_editor');
-    $this->createUserWithRole('all_editor', array('all_editor'));
+    $this->drupalCreateRole($perms, 'all_editor');
+    $this->config->set("roles.all_editor", 'perm')->save();
+    $this->createUserWithRole('all_editor', ['all_editor']);
 
-    // all_deletor
-    $perms = array(
+    // all_deletor.
+    $perms = [
       'access content',
-      _administerusersbyrole_build_perm_string(AccountInterface::AUTHENTICATED_ROLE, 'cancel'),
-    );
-    foreach ($this->roles as $roleName => $roleID) {
-      $perms[] = _administerusersbyrole_build_perm_string($roleID, 'cancel');
+      $this->accessManager->buildPermString('cancel'),
+    ];
+    foreach (array_keys($this->accessManager->managedRoles()) as $roleName) {
+      $perms[] = $this->accessManager->buildPermString('cancel', $roleName);
     }
-    $this->roles['all_deletor'] = $this->drupalCreateRole($perms, 'all_deletor');
-    $this->createUserWithRole('all_deletor', array('all_deletor'));
+    $this->drupalCreateRole($perms, 'all_deletor');
+    $this->createUserWithRole('all_deletor', ['all_deletor']);
 
-    // creator
-    $perms = array(
+    // Creator.
+    $perms = [
       'access content',
       'create users',
-    );
-    $this->roles['creator'] = $this->drupalCreateRole($perms, 'creator');
-    $this->createUserWithRole('creator', array('creator'));
+    ];
+    $this->drupalCreateRole($perms, 'creator');
+    $this->createUserWithRole('creator', ['creator']);
   }
 
+  /**
+   * Tests basic permissions.
+   */
   public function testPermissions() {
-    $expectations = array(
+    $expectations = [
       // When I'm logged in as...
-      'nobody' => array(
+      'nobody' => [
         // ...I can perform these actions on this other user...
-        'noroles'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'alpha'        => array('edit' => FALSE, 'cancel' => FALSE),
-        'alpha_editor' => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta'         => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta_editor'  => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta'    => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta_ed' => array('edit' => FALSE, 'cancel' => FALSE),
-        'creator'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_editor'   => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_deletor'  => array('edit' => FALSE, 'cancel' => FALSE),
+        'noroles'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'alpha'        => ['edit' => FALSE, 'cancel' => FALSE],
+        'alpha_editor' => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta'         => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta_editor'  => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta'    => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta_ed' => ['edit' => FALSE, 'cancel' => FALSE],
+        'creator'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_editor'   => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_deletor'  => ['edit' => FALSE, 'cancel' => FALSE],
         'create users' => FALSE,
-      ),
-      'noroles' => array(
-        'noroles'      => array('edit' => TRUE,  'cancel' => FALSE),
-        'alpha'        => array('edit' => FALSE, 'cancel' => FALSE),
-        'alpha_editor' => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta'         => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta_editor'  => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta'    => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta_ed' => array('edit' => FALSE, 'cancel' => FALSE),
-        'creator'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_editor'   => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_deletor'  => array('edit' => FALSE, 'cancel' => FALSE),
+      ],
+      'noroles' => [
+        'noroles'      => ['edit' => TRUE, 'cancel' => FALSE],
+        'alpha'        => ['edit' => FALSE, 'cancel' => FALSE],
+        'alpha_editor' => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta'         => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta_editor'  => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta'    => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta_ed' => ['edit' => FALSE, 'cancel' => FALSE],
+        'creator'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_editor'   => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_deletor'  => ['edit' => FALSE, 'cancel' => FALSE],
         'create users' => FALSE,
-      ),
-      'alpha' => array(
-        'noroles'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'alpha'        => array('edit' => TRUE,  'cancel' => FALSE),
-        'alpha_editor' => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta'         => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta_editor'  => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta'    => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta_ed' => array('edit' => FALSE, 'cancel' => FALSE),
-        'creator'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_editor'   => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_deletor'  => array('edit' => FALSE, 'cancel' => FALSE),
+      ],
+      'alpha' => [
+        'noroles'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'alpha'        => ['edit' => TRUE, 'cancel' => FALSE],
+        'alpha_editor' => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta'         => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta_editor'  => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta'    => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta_ed' => ['edit' => FALSE, 'cancel' => FALSE],
+        'creator'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_editor'   => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_deletor'  => ['edit' => FALSE, 'cancel' => FALSE],
         'create users' => FALSE,
-      ),
-      'alpha_editor' => array(
-        'noroles'      => array('edit' => TRUE,  'cancel' => FALSE),
-        'alpha'        => array('edit' => TRUE,  'cancel' => FALSE),
-        'alpha_editor' => array('edit' => TRUE,  'cancel' => FALSE),
-        'beta'         => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta_editor'  => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta'    => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta_ed' => array('edit' => FALSE, 'cancel' => FALSE),
-        'creator'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_editor'   => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_deletor'  => array('edit' => FALSE, 'cancel' => FALSE),
+      ],
+      'alpha_editor' => [
+        'noroles'      => ['edit' => TRUE, 'cancel' => FALSE],
+        'alpha'        => ['edit' => TRUE, 'cancel' => FALSE],
+        'alpha_editor' => ['edit' => TRUE, 'cancel' => FALSE],
+        'beta'         => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta_editor'  => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta'    => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta_ed' => ['edit' => FALSE, 'cancel' => FALSE],
+        'creator'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_editor'   => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_deletor'  => ['edit' => FALSE, 'cancel' => FALSE],
         'create users' => FALSE,
-      ),
-      'beta' => array(
-        'noroles'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'alpha'        => array('edit' => FALSE, 'cancel' => FALSE),
-        'alpha_editor' => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta'         => array('edit' => TRUE,  'cancel' => FALSE),
-        'beta_editor'  => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta'    => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta_ed' => array('edit' => FALSE, 'cancel' => FALSE),
-        'creator'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_editor'   => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_deletor'  => array('edit' => FALSE, 'cancel' => FALSE),
+      ],
+      'beta' => [
+        'noroles'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'alpha'        => ['edit' => FALSE, 'cancel' => FALSE],
+        'alpha_editor' => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta'         => ['edit' => TRUE, 'cancel' => FALSE],
+        'beta_editor'  => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta'    => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta_ed' => ['edit' => FALSE, 'cancel' => FALSE],
+        'creator'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_editor'   => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_deletor'  => ['edit' => FALSE, 'cancel' => FALSE],
         'create users' => FALSE,
-      ),
-      'beta_editor' => array(
-        'noroles'      => array('edit' => TRUE,  'cancel' => FALSE),
-        'alpha'        => array('edit' => FALSE, 'cancel' => FALSE),
-        'alpha_editor' => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta'         => array('edit' => TRUE,  'cancel' => TRUE),
-        'beta_editor'  => array('edit' => TRUE,  'cancel' => FALSE),
-        'alphabeta'    => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta_ed' => array('edit' => FALSE, 'cancel' => FALSE),
-        'creator'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_editor'   => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_deletor'  => array('edit' => FALSE, 'cancel' => FALSE),
+      ],
+      'beta_editor' => [
+        'noroles'      => ['edit' => TRUE, 'cancel' => TRUE],
+        'alpha'        => ['edit' => FALSE, 'cancel' => FALSE],
+        'alpha_editor' => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta'         => ['edit' => TRUE, 'cancel' => TRUE],
+        'beta_editor'  => ['edit' => TRUE, 'cancel' => FALSE],
+        'alphabeta'    => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta_ed' => ['edit' => FALSE, 'cancel' => FALSE],
+        'creator'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_editor'   => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_deletor'  => ['edit' => FALSE, 'cancel' => FALSE],
         'create users' => FALSE,
-      ),
-      'alphabeta' => array(
-        'noroles'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'alpha'        => array('edit' => FALSE, 'cancel' => FALSE),
-        'alpha_editor' => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta'         => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta_editor'  => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta'    => array('edit' => TRUE,  'cancel' => FALSE),
-        'alphabeta_ed' => array('edit' => FALSE, 'cancel' => FALSE),
-        'creator'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_editor'   => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_deletor'  => array('edit' => FALSE, 'cancel' => FALSE),
+      ],
+      'alphabeta' => [
+        'noroles'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'alpha'        => ['edit' => FALSE, 'cancel' => FALSE],
+        'alpha_editor' => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta'         => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta_editor'  => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta'    => ['edit' => TRUE, 'cancel' => FALSE],
+        'alphabeta_ed' => ['edit' => FALSE, 'cancel' => FALSE],
+        'creator'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_editor'   => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_deletor'  => ['edit' => FALSE, 'cancel' => FALSE],
         'create users' => FALSE,
-      ),
-      'alphabeta_ed' => array(
-        'noroles'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'alpha'        => array('edit' => TRUE,  'cancel' => FALSE),
-        'alpha_editor' => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta'         => array('edit' => TRUE,  'cancel' => FALSE),
-        'beta_editor'  => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta'    => array('edit' => TRUE,  'cancel' => FALSE),
-        'alphabeta_ed' => array('edit' => TRUE,  'cancel' => FALSE),
-        'creator'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_editor'   => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_deletor'  => array('edit' => FALSE, 'cancel' => FALSE),
+      ],
+      'alphabeta_ed' => [
+        'noroles'      => ['edit' => TRUE, 'cancel' => FALSE],
+        'alpha'        => ['edit' => TRUE, 'cancel' => FALSE],
+        'alpha_editor' => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta'         => ['edit' => TRUE, 'cancel' => FALSE],
+        'beta_editor'  => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta'    => ['edit' => TRUE, 'cancel' => FALSE],
+        'alphabeta_ed' => ['edit' => TRUE, 'cancel' => FALSE],
+        'creator'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_editor'   => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_deletor'  => ['edit' => FALSE, 'cancel' => FALSE],
         'create users' => FALSE,
-      ),
-      'all_editor' => array(
-        'noroles'      => array('edit' => TRUE,  'cancel' => FALSE),
-        'alpha'        => array('edit' => TRUE,  'cancel' => FALSE),
-        'alpha_editor' => array('edit' => TRUE,  'cancel' => FALSE),
-        'beta'         => array('edit' => TRUE,  'cancel' => FALSE),
-        'beta_editor'  => array('edit' => TRUE,  'cancel' => FALSE),
-        'alphabeta'    => array('edit' => TRUE,  'cancel' => FALSE),
-        'alphabeta_ed' => array('edit' => TRUE,  'cancel' => FALSE),
-        'creator'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_editor'   => array('edit' => TRUE,  'cancel' => FALSE),
-        'all_deletor'  => array('edit' => FALSE, 'cancel' => FALSE),
+      ],
+      'all_editor' => [
+        'noroles'      => ['edit' => TRUE, 'cancel' => FALSE],
+        'alpha'        => ['edit' => TRUE, 'cancel' => FALSE],
+        'alpha_editor' => ['edit' => TRUE, 'cancel' => FALSE],
+        'beta'         => ['edit' => TRUE, 'cancel' => FALSE],
+        'beta_editor'  => ['edit' => TRUE, 'cancel' => FALSE],
+        'alphabeta'    => ['edit' => TRUE, 'cancel' => FALSE],
+        'alphabeta_ed' => ['edit' => TRUE, 'cancel' => FALSE],
+        'creator'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_editor'   => ['edit' => TRUE, 'cancel' => FALSE],
+        'all_deletor'  => ['edit' => FALSE, 'cancel' => FALSE],
         'create users' => FALSE,
-      ),
-      'all_deletor' => array(
-        'noroles'      => array('edit' => FALSE, 'cancel' => TRUE),
-        'alpha'        => array('edit' => FALSE, 'cancel' => TRUE),
-        'alpha_editor' => array('edit' => FALSE, 'cancel' => TRUE),
-        'beta'         => array('edit' => FALSE, 'cancel' => TRUE),
-        'beta_editor'  => array('edit' => FALSE, 'cancel' => TRUE),
-        'alphabeta'    => array('edit' => FALSE, 'cancel' => TRUE),
-        'alphabeta_ed' => array('edit' => FALSE, 'cancel' => TRUE),
-        'creator'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_editor'   => array('edit' => FALSE, 'cancel' => TRUE),
-        'all_deletor'  => array('edit' => TRUE,  'cancel' => FALSE),
+      ],
+      'all_deletor' => [
+        'noroles'      => ['edit' => FALSE, 'cancel' => TRUE],
+        'alpha'        => ['edit' => FALSE, 'cancel' => TRUE],
+        'alpha_editor' => ['edit' => FALSE, 'cancel' => TRUE],
+        'beta'         => ['edit' => FALSE, 'cancel' => TRUE],
+        'beta_editor'  => ['edit' => FALSE, 'cancel' => TRUE],
+        'alphabeta'    => ['edit' => FALSE, 'cancel' => TRUE],
+        'alphabeta_ed' => ['edit' => FALSE, 'cancel' => TRUE],
+        'creator'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_editor'   => ['edit' => FALSE, 'cancel' => TRUE],
+        'all_deletor'  => ['edit' => TRUE, 'cancel' => FALSE],
         'create users' => FALSE,
-      ),
-      'creator' => array(
-        'noroles'      => array('edit' => FALSE, 'cancel' => FALSE),
-        'alpha'        => array('edit' => FALSE, 'cancel' => FALSE),
-        'alpha_editor' => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta'         => array('edit' => FALSE, 'cancel' => FALSE),
-        'beta_editor'  => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta'    => array('edit' => FALSE, 'cancel' => FALSE),
-        'alphabeta_ed' => array('edit' => FALSE, 'cancel' => FALSE),
-        'creator'      => array('edit' => TRUE,  'cancel' => FALSE),
-        'all_editor'   => array('edit' => FALSE, 'cancel' => FALSE),
-        'all_deletor'  => array('edit' => FALSE, 'cancel' => FALSE),
+      ],
+      'creator' => [
+        'noroles'      => ['edit' => FALSE, 'cancel' => FALSE],
+        'alpha'        => ['edit' => FALSE, 'cancel' => FALSE],
+        'alpha_editor' => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta'         => ['edit' => FALSE, 'cancel' => FALSE],
+        'beta_editor'  => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta'    => ['edit' => FALSE, 'cancel' => FALSE],
+        'alphabeta_ed' => ['edit' => FALSE, 'cancel' => FALSE],
+        'creator'      => ['edit' => TRUE, 'cancel' => FALSE],
+        'all_editor'   => ['edit' => FALSE, 'cancel' => FALSE],
+        'all_deletor'  => ['edit' => FALSE, 'cancel' => FALSE],
         'create users' => TRUE,
-      ),
-    );
+      ],
+    ];
 
+    $assert = $this->assertSession();
     foreach ($expectations as $loginUsername => $editUsernames) {
       if ($loginUsername !== 'nobody') {
         $this->drupalLogin($this->users[$loginUsername]);
@@ -237,20 +269,19 @@ class AdministerusersbyroleTest extends BrowserTestBase {
           $editUid = $this->users[$editUsername]->id();
           foreach ($operations as $operation => $expectedResult) {
             $this->drupalGet("user/$editUid/$operation");
-            // $loginUsername perform $operation on $editUsername
+
             if ($expectedResult) {
               if ($operation === 'edit') {
-                $this->assertRaw("All emails from the system will be sent to this address.");
+                $assert->responseContains("All emails from the system will be sent to this address.");
               }
               elseif ($operation === 'cancel') {
-                $this->assertRaw("Are you sure you want to cancel the account <em class=\"placeholder\">$editUsername</em>?");
+                $assert->responseContains("Are you sure you want to cancel the account <em class=\"placeholder\">$editUsername</em>?");
               }
             }
             else {
-              $this->assertTrue(
-                strstr($this->getRawContent(), "You do not have permission to $operation <em class=\"placeholder\">$editUsername</em>.")
-                || strstr($this->getRawContent(), 'Access denied'),
-                "My expectation is that $loginUsername shouldn't be able to $operation $editUsername, but it can.");
+              $content = $this->getSession()->getPage()->getContent();
+              $denied = strstr($content, "You do not have permission to $operation <em class=\"placeholder\">$editUsername</em>.") || strstr($content, 'Access denied');
+              $this->assertTrue($denied, "My expectation is that $loginUsername shouldn't be able to $operation $editUsername, but it can.");
             }
           }
         }
@@ -262,9 +293,12 @@ class AdministerusersbyroleTest extends BrowserTestBase {
     }
   }
 
+  /**
+   * Creates a user with the specified name and roles.
+   */
   protected function createUserWithRole($userName, $roleNames) {
     $user = $this->drupalCreateUser([], $userName);
-    $this->assertTrue($user, "Unable to create user $userName.");
+    $this->assertNotEmpty($user, "Unable to create user $userName.");
     foreach ($roleNames as $role) {
       $user->addRole($role);
     }
@@ -272,23 +306,29 @@ class AdministerusersbyroleTest extends BrowserTestBase {
     $this->users[$userName] = $user;
   }
 
+  /**
+   * Creates and role, a user with that roles and a user that can edit the
+   * role.
+   */
   protected function createRolesAndUsers($roleName, $allowEditorToCancel) {
-    // create basic role
-    $this->roles[$roleName] = $this->drupalCreateRole(array('access content'), $roleName);
-    $this->createUserWithRole($roleName, array($roleName));
+    // Create basic role.
+    $this->drupalCreateRole(['access content'], $roleName);
+    $this->config->set("roles.$roleName", 'perm')->save();
+    $this->createUserWithRole($roleName, [$roleName]);
 
-    // create role to edit above role and also anyone with no custom roles.
-    $perms = array(
+    // Create role to edit above role and also anyone with no custom roles.
+    $perms = [
       'access content',
-      _administerusersbyrole_build_perm_string(AccountInterface::AUTHENTICATED_ROLE, 'edit'),
-      _administerusersbyrole_build_perm_string($this->roles[$roleName], 'edit'),
-    );
+      $this->accessManager->buildPermString('edit'),
+      $this->accessManager->buildPermString('edit', $roleName),
+    ];
     if ($allowEditorToCancel) {
-      // Don't add in "no custom roles" this time, to give better variety of testing.
-      $perms[] = _administerusersbyrole_build_perm_string($this->roles[$roleName], 'cancel');
+      $perms[] = $this->accessManager->buildPermString('cancel');
+      $perms[] = $this->accessManager->buildPermString('cancel', $roleName);
     }
-    $this->roles["{$roleName}_editor"] = $this->drupalCreateRole($perms, "{$roleName}_editor");
-    $this->createUserWithRole("{$roleName}_editor", array("{$roleName}_editor"));
+    $this->drupalCreateRole($perms, "{$roleName}_editor");
+    $this->config->set("roles.{$roleName}_editor", 'perm')->save();
+    $this->createUserWithRole("{$roleName}_editor", ["{$roleName}_editor"]);
   }
 
 }