+++ modules/contrib/eck/tests/src/Functional/ConfigImportTest.php
@@ -3,10 +3,9 @@
 namespace Drupal\Tests\eck\Functional;
 
 use Drupal\Core\Url;
-use Drupal\simpletest\WebTestBase;
 
 /**
- * Class ConfigImportTest
+ * Class ConfigImportTest.
  *
  * @group eck
  */
@@ -43,6 +42,9 @@ class ConfigImportTest extends FunctionalTestBase {
     }
   }
 
+  /**
+   * Tests the import of configuration.
+   */
   public function testImport() {
     $defaultLanguage = \Drupal::languageManager()->getDefaultLanguage();
 
@@ -85,7 +87,7 @@ class ConfigImportTest extends FunctionalTestBase {
     $sync->write($entityBundleConfigName, $entityBundleConfiguration);
 
     // Import the configuration.
-    $this->drupalPostForm(Url::fromRoute('config.sync'), [], t('Import all'));
+    $this->drupalPostForm(Url::fromRoute('config.sync'), [], $this->t('Import all'));
 
     // Verify the values appeared.
     $config = $this->config($entityTypeConfigName);
@@ -103,9 +105,9 @@ class ConfigImportTest extends FunctionalTestBase {
     $route = 'eck.entity.add';
     $routeArguments = [
       'eck_entity_type' => 'test_entity',
-      'eck_entity_bundle' => 'bundle'
+      'eck_entity_bundle' => 'bundle',
     ];
-    $this->drupalPostForm(Url::fromRoute($route, $routeArguments), $edit, t('Save'));
+    $this->drupalPostForm(Url::fromRoute($route, $routeArguments), $edit, $this->t('Save'));
     $this->assertSession()->responseContains($edit['title[0][value]']);
   }